Add test case for logic function with variant
