+++ creusot-contracts/src/std.rs
@@ -1,3 +1,5 @@
 pub mod vec;
+pub mod well_founded;
 
 pub use vec::*;
+pub use well_founded::*;
