+++ creusot/tests/should_succeed/cell/02.stdout
@@ -169,14 +169,14 @@ module C02_LemmaFibBound_Impl
         end
       end
 end
-module C02_Impl2_Inv_Interface
+module C02_Impl1_Inv_Interface
   use prelude.Prelude
   use Type
   use mach.int.Int
   use mach.int.UInt64
   predicate inv (self : Type.c02_fib) (v : Type.core_option_option usize)
 end
-module C02_Impl2_Inv
+module C02_Impl1_Inv
   use prelude.Prelude
   use Type
   use mach.int.Int
@@ -188,13 +188,13 @@ module C02_Impl2_Inv
       | Type.Core_Option_Option_Some i -> UInt64.to_int i = Fib0.fib (UInt64.to_int (let Type.C02_Fib a = self in a))
       end
 end
-module C02_Impl2
+module C02_Impl1
   use Type
   use mach.int.Int
   use prelude.Prelude
   use mach.int.UInt64
   clone C02_Fib as Fib0 with axiom .
-  clone C02_Impl2_Inv as Inv0 with function Fib0.fib = Fib0.fib
+  clone C02_Impl1_Inv as Inv0 with function Fib0.fib = Fib0.fib
   clone C02_Inv_Inv as Inv1 with type self = Type.c02_fib, type t = Type.core_option_option usize,
   predicate inv = Inv0.inv
 end
@@ -438,7 +438,7 @@ module C02_FibMemo
   clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = usize
   clone CreusotContracts_Std1_Vec_Impl3_Index_Interface as Index0 with type t = Type.c02_cell (Type.core_option_option usize) (Type.c02_fib),
   function Model0.model = Model0.model
-  clone C02_Impl2_Inv as Inv0 with function Fib0.fib = Fib0.fib
+  clone C02_Impl1_Inv as Inv0 with function Fib0.fib = Fib0.fib
   clone C02_Impl0_Set_Interface as Set0 with type t = Type.core_option_option usize, type i = Type.c02_fib,
   predicate Inv0.inv = Inv0.inv
   clone C02_Impl0_Get_Interface as Get0 with type t = Type.core_option_option usize, type i = Type.c02_fib,
@@ -613,8 +613,5 @@ module C02_FibMemo
   BB17 {
     return _0
   }
-  
-end
-module C02_Impl1
   
 end
