+++ creusot/tests/should_succeed/inc_some_2_tree.stdout
@@ -224,27 +224,27 @@ module CreusotContracts_Builtins_Model_Impl1
   clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy1 with type self = borrowed t,
   type modelty = ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Int_Impl4_ModelTy
+module CreusotContracts_Builtins_Int_Impl5_ModelTy
   use mach.int.Int
   type modelty  = 
     int
 end
-module CreusotContracts_Builtins_Int_Impl4_Model_Interface
+module CreusotContracts_Builtins_Int_Impl5_Model_Interface
   use mach.int.Int
   use mach.int.UInt32
   function model (self : uint32) : int
 end
-module CreusotContracts_Builtins_Int_Impl4_Model
+module CreusotContracts_Builtins_Int_Impl5_Model
   use mach.int.Int
   use mach.int.UInt32
   function model (self : uint32) : int = 
     UInt32.to_int self
 end
-module CreusotContracts_Builtins_Int_Impl4
+module CreusotContracts_Builtins_Int_Impl5
   use mach.int.Int
   use mach.int.UInt32
-  clone CreusotContracts_Builtins_Int_Impl4_Model as Model0
-  clone CreusotContracts_Builtins_Int_Impl4_ModelTy as ModelTy0
+  clone CreusotContracts_Builtins_Int_Impl5_Model as Model0
+  clone CreusotContracts_Builtins_Int_Impl5_ModelTy as ModelTy0
   clone CreusotContracts_Builtins_Model_Model_Model as Model1 with type self = uint32,
   type ModelTy0.modelty = ModelTy0.modelty, function model = Model0.model
   clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy1 with type self = uint32,
@@ -285,7 +285,7 @@ module IncSome2Tree_Impl1_TakeSomeRest_Interface
   use prelude.Prelude
   use Type
   use mach.int.Int
-  clone CreusotContracts_Builtins_Int_Impl4_ModelTy as ModelTy0
+  clone CreusotContracts_Builtins_Int_Impl5_ModelTy as ModelTy0
   clone CreusotContracts_Builtins_Model_Impl1_Model_Interface as Model0 with type t = uint32,
   type ModelTy0.modelty = ModelTy0.modelty
   clone IncSome2Tree_Impl1_Sum_Interface as Sum0
@@ -301,8 +301,8 @@ module IncSome2Tree_Impl1_TakeSomeRest
   use prelude.Prelude
   use Type
   use mach.int.Int
-  clone CreusotContracts_Builtins_Int_Impl4_Model as Model1
-  clone CreusotContracts_Builtins_Int_Impl4_ModelTy as ModelTy0
+  clone CreusotContracts_Builtins_Int_Impl5_Model as Model1
+  clone CreusotContracts_Builtins_Int_Impl5_ModelTy as ModelTy0
   clone CreusotContracts_Builtins_Model_Impl1_Model as Model0 with type t = uint32,
   type ModelTy0.modelty = ModelTy0.modelty, function Model0.model = Model1.model
   clone IncSome2Tree_Impl1_Sum as Sum0
@@ -519,7 +519,7 @@ module IncSome2Tree_IncSome2Tree
   use Type
   clone IncSome2Tree_Impl1_Sum as Sum0
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Int_Impl4_Model as Model1
+  clone CreusotContracts_Builtins_Int_Impl5_Model as Model1
   clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve5 with type self = Type.incsome2tree_tree
   clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve4 with type self = ()
   clone Core_Panicking_Panic_Interface as Panic0
@@ -529,7 +529,7 @@ module IncSome2Tree_IncSome2Tree
   clone CreusotContracts_Builtins_Resolve_Impl0_Resolve as Resolve2 with type t1 = borrowed uint32,
   type t2 = borrowed (Type.incsome2tree_tree), predicate Resolve0.resolve = Resolve0.resolve,
   predicate Resolve1.resolve = Resolve1.resolve
-  clone CreusotContracts_Builtins_Int_Impl4_ModelTy as ModelTy0
+  clone CreusotContracts_Builtins_Int_Impl5_ModelTy as ModelTy0
   clone CreusotContracts_Builtins_Model_Impl1_Model as Model0 with type t = uint32,
   type ModelTy0.modelty = ModelTy0.modelty, function Model0.model = Model1.model
   clone IncSome2Tree_Impl1_TakeSomeRest_Interface as TakeSomeRest0 with function Sum0.sum = Sum0.sum,
