+++ creusot/tests/should_succeed/inc_some_tree.stdout
@@ -224,27 +224,27 @@ module CreusotContracts_Builtins_Model_Impl1
   clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy1 with type self = borrowed t,
   type modelty = ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Int_Impl4_ModelTy
+module CreusotContracts_Builtins_Int_Impl5_ModelTy
   use mach.int.Int
   type modelty  = 
     int
 end
-module CreusotContracts_Builtins_Int_Impl4_Model_Interface
+module CreusotContracts_Builtins_Int_Impl5_Model_Interface
   use mach.int.Int
   use mach.int.UInt32
   function model (self : uint32) : int
 end
-module CreusotContracts_Builtins_Int_Impl4_Model
+module CreusotContracts_Builtins_Int_Impl5_Model
   use mach.int.Int
   use mach.int.UInt32
   function model (self : uint32) : int = 
     UInt32.to_int self
 end
-module CreusotContracts_Builtins_Int_Impl4
+module CreusotContracts_Builtins_Int_Impl5
   use mach.int.Int
   use mach.int.UInt32
-  clone CreusotContracts_Builtins_Int_Impl4_Model as Model0
-  clone CreusotContracts_Builtins_Int_Impl4_ModelTy as ModelTy0
+  clone CreusotContracts_Builtins_Int_Impl5_Model as Model0
+  clone CreusotContracts_Builtins_Int_Impl5_ModelTy as ModelTy0
   clone CreusotContracts_Builtins_Model_Model_Model as Model1 with type self = uint32,
   type ModelTy0.modelty = ModelTy0.modelty, function model = Model0.model
   clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy1 with type self = uint32,
@@ -286,7 +286,7 @@ module IncSomeTree_Impl1_TakeSome_Interface
   use Type
   use mach.int.Int
   clone IncSomeTree_Impl1_Sum_Interface as Sum0
-  clone CreusotContracts_Builtins_Int_Impl4_ModelTy as ModelTy0
+  clone CreusotContracts_Builtins_Int_Impl5_ModelTy as ModelTy0
   clone CreusotContracts_Builtins_Model_Impl1_Model_Interface as Model0 with type t = uint32,
   type ModelTy0.modelty = ModelTy0.modelty
   val take_some (self : borrowed (Type.incsometree_tree)) : borrowed uint32
@@ -300,8 +300,8 @@ module IncSomeTree_Impl1_TakeSome
   use prelude.Prelude
   use Type
   use mach.int.Int
-  clone CreusotContracts_Builtins_Int_Impl4_Model as Model1
-  clone CreusotContracts_Builtins_Int_Impl4_ModelTy as ModelTy0
+  clone CreusotContracts_Builtins_Int_Impl5_Model as Model1
+  clone CreusotContracts_Builtins_Int_Impl5_ModelTy as ModelTy0
   clone CreusotContracts_Builtins_Model_Impl1_Model as Model0 with type t = uint32,
   type ModelTy0.modelty = ModelTy0.modelty, function Model0.model = Model1.model
   clone IncSomeTree_Impl1_Sum as Sum0
@@ -492,13 +492,13 @@ module IncSomeTree_IncSomeTree
   use Type
   clone IncSomeTree_Impl1_Sum as Sum0
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Int_Impl4_Model as Model1
+  clone CreusotContracts_Builtins_Int_Impl5_Model as Model1
   clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve3 with type self = Type.incsometree_tree
   clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = ()
   clone Core_Panicking_Panic_Interface as Panic0
   clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve1 with type t = uint32
   clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
-  clone CreusotContracts_Builtins_Int_Impl4_ModelTy as ModelTy0
+  clone CreusotContracts_Builtins_Int_Impl5_ModelTy as ModelTy0
   clone CreusotContracts_Builtins_Model_Impl1_Model as Model0 with type t = uint32,
   type ModelTy0.modelty = ModelTy0.modelty, function Model0.model = Model1.model
   clone IncSomeTree_Impl1_TakeSome_Interface as TakeSome0 with function Model0.model = Model0.model,
